requirements.txt
django
requests
bs4
django-mail-queue